# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Flakey API
  version: 0.1.0
  description: |
    HTTP API for the Flakey test-reporting dashboard.

    **This is the source of truth for the API contract.** `frontend/src/lib/api-generated.ts`
    is generated from it via `pnpm openapi:generate` (openapi-typescript). When you
    add or change a route, update this spec **in the same commit** — the
    `pnpm openapi:check` drift check (and CI) flags a stale spec.

    Scope: this is a **seed** covering the core, integrator-facing surface (runs,
    tests, errors, stats, flaky, compare, auth, health) plus the two AI write
    routes that open repo PRs / cluster failures (`/analyze/clusters`,
    `/analyze/fix-pr`). Remaining routes (other `/analyze/*`, integrations,
    releases, manual tests, live SSE, uploads, quarantine, views, coverage,
    support) are TODO — add them here as the contract is migrated off the
    hand-synced `frontend/src/lib/api.ts`. Run `pnpm endpoint-inventory` (the
    skill) for the full route list to work through.

    Auth: every endpoint except `/health`, `/auth/login`, and `/auth/register`
    requires a Bearer token — a JWT from login, or a `fk_…` API key. All
    tenant-scoped responses are filtered to the caller's org by Postgres RLS
    (not expressible in OpenAPI; see docs/architecture.md §4).
servers:
  - url: "{baseUrl}"
    description: Backend API origin (FLAKEY_API_URL / VITE_API_URL)
    variables:
      baseUrl:
        default: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /errors:
    get:
      tags: [errors]
      summary: Recurring failures grouped by error fingerprint
      description: |
        Returns the 100 most-recent error groups (by last_seen, then
        occurrence_count). The `status` filter is applied **before** that cap,
        so a status rarer than the 100 newest groups is not truncated. An
        unrecognised `status` value is ignored (no filter applied).
      parameters:
        - { name: suite, in: query, schema: { type: string } }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ErrorStatus" }
      responses:
        "200":
          description: Error groups
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ErrorGroup" }
  /errors/{fingerprint}/status:
    parameters:
      - { name: fingerprint, in: path, required: true, schema: { type: string } }
    patch:
      tags: [errors]
      summary: Set the triage status of an error group (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/ErrorStatus" }
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: boolean }
                  group_id: { type: integer }
                  status: { $ref: "#/components/schemas/ErrorStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /errors/{fingerprint}/assign:
    parameters:
      - { name: fingerprint, in: path, required: true, schema: { type: string } }
    post:
      tags: [errors]
      summary: Assign (or un-assign) an owner to an error group (upsert)
      description: |
        Lightweight triage ownership — "who's chasing this failure". Pass
        `user_id: null` to un-assign. The user must be a member of the caller's
        org. Viewer role is rejected (403).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: [integer, "null"] }
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned: { type: boolean }
                  group_id: { type: integer }
                  user_id: { type: [integer, "null"] }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: Viewer role — assigning an error is admin+
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /errors/{fingerprint}/tests:
    parameters:
      - { name: fingerprint, in: path, required: true, schema: { type: string } }
    get:
      tags: [errors]
      summary: Tests affected by an error group
      responses:
        "200":
          description: Affected tests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AffectedTest" }
  /errors/{fingerprint}/notes:
    parameters:
      - { name: fingerprint, in: path, required: true, schema: { type: string } }
    get:
      tags: [errors]
      summary: Notes on an error group
      responses:
        "200":
          description: Notes (oldest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ErrorNote" }
    post:
      tags: [errors]
      summary: Add a note to an error group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorNote" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A login JWT or an `fk_…` API key.
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    ErrorStatus:
      type: string
      description: Triage state of an error group. Unstamped groups default to `open`.
      enum: [open, investigating, known, fixed, ignored]
    ErrorGroup:
      type: object
      properties:
        fingerprint: { type: string }
        error_message: { type: string }
        occurrence_count: { type: integer }
        affected_tests: { type: integer }
        affected_runs: { type: integer }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        latest_run_id: { type: integer }
        latest_test_id: { type: [integer, "null"] }
        test_titles: { type: array, items: { type: string } }
        file_paths: { type: array, items: { type: string } }
        suite_name: { type: string }
        group_id: { type: [integer, "null"] }
        status: { $ref: "#/components/schemas/ErrorStatus" }
        assigned_to: { type: [integer, "null"] }
        assigned_to_email: { type: [string, "null"] }
        note_count: { type: integer }
    AffectedTest:
      type: object
      properties:
        full_title: { type: string }
        title: { type: string }
        file_path: { type: string }
        suite_name: { type: string }
        occurrence_count: { type: integer }
        last_seen: { type: string, format: date-time }
        latest_test_id: { type: integer }
        latest_run_id: { type: integer }
    ErrorNote:
      type: object
      properties:
        id: { type: integer }
        body: { type: string }
        created_at: { type: string, format: date-time }
        user_name: { type: [string, "null"] }
        user_email: { type: [string, "null"] }
...